+++ lib/active_delivery.rb
@@ -13,3 +13,6 @@ require "active_delivery/lines/mailer" if defined?(ActionMailer)
 
 require "active_delivery/raitie" if defined?(::Rails::Railtie)
 require "active_delivery/testing" if ENV["RACK_ENV"] == "test" || ENV["RAILS_ENV"] == "test"
+
+require "abstract_notifier"
+require "active_delivery/lines/notifier"